feat: добавлен REST API для получения города по имени
- Реализован CityRepository с методом findByCityName()
- Создан CityService как промежуточный слой (пока без бизнес-логики)
- Добавлен CityController с endpoint GET /api/cities/{name}
- Настроена инъекция зависимостей между слоями через конструкторы

+++ src/main/resources/db/01_create_cities_table.sql
@@ -1,5 +1,6 @@
-CREATE TABLE IF NOT EXISTS cities (
+CREATE TABLE IF NOT EXISTS cities
+(
     id          BIGSERIAL PRIMARY KEY,
-    city_name   VARCHAR(20) NOT NULL UNIQUE,
-    temperature DECIMAL     NOT NULL
+    city_name   VARCHAR(20)      NOT NULL UNIQUE,
+    temperature DOUBLE PRECISION NOT NULL
 );
